Refactored for CSV Readers to only be on one file
