+++ projectApp/src/models/CSVDataObject.tsx
@@ -1,6 +1,6 @@
-import { LocalCSVReader, LocalCsvReader } from "../components/CSV_Readers/LocalCSVReader";
+import { LocalCSVReader, LocalCsvReader, UrlCSVReader } from "../components/CSV_Readers/CSVReaders";
 import { CSVDataInterface } from "../types/CSVInterfaces";
-import { UrlCSVReader } from "../components/CSV_Readers/UrlCSVReader";
+
 
 export class CSVDataObject implements CSVDataInterface{
     name: string;
