+++ projectApp/test/csvTesting.test.ts
@@ -1,8 +1,7 @@
 import { describe } from 'node:test';
 import { expect, test } from 'vitest';
 
-import { LocalCSVReader as localReader, LocalCSVHeaders as localHeaders } from "../src/components/CSV_Readers/LocalCSVReader.tsx";
-import { UrlCSVReader as urlReader, UrlCSVHeaders as urlHeaders } from "../src/components/CSV_Readers/UrlCSVReader.tsx";
+import { LocalCSVReader as localReader, LocalCSVHeaders as localHeaders, UrlCSVReader as urlReader, UrlCSVHeaders as urlHeaders } from "../src/components/CSV_Readers/CSVReaders.tsx";
 import { CSVHeaders, TimeSeriesData } from '../src/types/CSVInterfaces.tsx';
